Rename views to components, use .js extensions, instantiatable console
