+++ src/components/App.test.js
@@ -1,4 +1,5 @@
 import assert from 'assert';
+import App from './App';
 
 describe('App', function() {
   describe('#setContent()', function() {
