+++ src/components/content/console/Console.test.js
@@ -1,9 +1,11 @@
+import jsx from '../../../utilities/jsx';
 import assert from 'assert';
-import * as console from './Console';
+import Console from './Console';
 import { stdout, stderr } from 'test-console';
 
 describe('Console', function() {
   before(function() {
+    this.console = Console();
     this.oldWindowOnError = window.onerror && window.onerror.bind({});
     this.oldWindowConsole = {
       ...window.console
@@ -12,16 +14,16 @@ describe('Console', function() {
 
   describe('#setHistory()', function() {
     it('should store history in reverse order', function() {
-      console.setHistory("'history1'");
-      console.setHistory("'history2'");
-      console.setHistory("'history3'");
-      assert.deepStrictEqual(console.getHistory(), ["'history3'", "'history2'", "'history1'"]);
+      this.console.setHistory("'history1'");
+      this.console.setHistory("'history2'");
+      this.console.setHistory("'history3'");
+      assert.deepStrictEqual(this.console.getHistory(), ["'history3'", "'history2'", "'history1'"]);
     });
     it('should store the most recent 50 commands', function() {
       for (let i = 1; i <= 60; i++) {
-        console.setHistory(`'history${i}'`);
+        this.console.setHistory(`'history${i}'`);
       }
-      const history = console.getHistory();
+      const history = this.console.getHistory();
       assert.strictEqual(history.length, 50);
       assert.deepStrictEqual(history[0], "'history60'");
       assert.deepStrictEqual(history[49], "'history11'");
@@ -30,19 +32,19 @@ describe('Console', function() {
 
   describe('#getHistory()', function() {
     it('should retrieve the history', function() {
-      console.setHistory(null, 0);
-      console.setHistory("'history'");
-      const history = console.getHistory();
+      this.console.setHistory(null, 0);
+      this.console.setHistory("'history'");
+      const history = this.console.getHistory();
       assert.deepStrictEqual(history, ["'history'"]);
-      console.setHistory("'history2'");
-      const history2 = console.getHistory();
+      this.console.setHistory("'history2'");
+      const history2 = this.console.getHistory();
       assert.deepStrictEqual(history2, ["'history2'", "'history'"]);
     });
   });
 
   describe('#parseStack()', function() {
     it('should parse stack with backslashes', function() {
-      const parsed = console.parseStack('\n\n  at fn (C:\\path\\file.js:123:45)');
+      const parsed = this.console.parseStack('\n\n  at fn (C:\\path\\file.js:123:45)');
       assert.deepStrictEqual(parsed[0].path, 'fn (C:\\path\\file.js:123:45)');
       assert.deepStrictEqual(parsed[0].url, undefined);
       assert.deepStrictEqual(parsed[0].fileName, 'file.js');
@@ -50,7 +52,7 @@ describe('Console', function() {
       assert.deepStrictEqual(parsed[0].columnNumber, '45');
     });
     it('should parse stack with forwardslashes', function() {
-      const parsed = console.parseStack('\n\n  at fn (http://path/file.js:123:45)');
+      const parsed = this.console.parseStack('\n\n  at fn (http://path/file.js:123:45)');
       assert.deepStrictEqual(parsed[0].path, 'fn (http://path/file.js:123:45)');
       assert.deepStrictEqual(parsed[0].url, 'http://path/file.js');
       assert.deepStrictEqual(parsed[0].fileName, 'file.js');
@@ -58,7 +60,7 @@ describe('Console', function() {
       assert.deepStrictEqual(parsed[0].columnNumber, '45');
     });
     it('should parse stack with no parenthesis', function() {
-      const parsed = console.parseStack('\n\n  at C:/path/file.js:123:45');
+      const parsed = this.console.parseStack('\n\n  at C:/path/file.js:123:45');
       assert.deepStrictEqual(parsed[0].path, 'C:/path/file.js:123:45');
       assert.deepStrictEqual(parsed[0].url, undefined);
       assert.deepStrictEqual(parsed[0].fileName, 'file.js');
@@ -66,21 +68,21 @@ describe('Console', function() {
       assert.deepStrictEqual(parsed[0].columnNumber, '45');
     });
     it('should handle unexpected input', function() {
-      const parsed = console.parseStack('\n\n something that does not look like a stack');
+      const parsed = this.console.parseStack('\n\n something that does not look like a stack');
       assert.deepStrictEqual(parsed, []);
     });
   });
 
   describe('#overrideWindowConsole()', function() {
     it('should create a new function for the 4 verb methods', function() {
-      console.overrideWindowConsole();
+      this.console.overrideWindowConsole();
       assert.notDeepStrictEqual(this.oldWindowConsole.error, window.console.error);
       assert.notDeepStrictEqual(this.oldWindowConsole.info, window.console.info);
       assert.notDeepStrictEqual(this.oldWindowConsole.log, window.console.log);
       assert.notDeepStrictEqual(this.oldWindowConsole.warn, window.console.warn);
     });
     it('should not break the standard console', function() {
-      console.overrideWindowConsole();
+      this.console.overrideWindowConsole();
       stdout.inspectSync((output) => {
         window.console.log('string');
         window.console.log([1, 2, 3]);
@@ -92,20 +94,20 @@ describe('Console', function() {
 
   describe('#overrideWindowOnError()', function() {
     it('should create a new function for window.onerror', function() {
-      console.restoreWindowOnError();
-      console.overrideWindowOnError();
+      this.console.restoreWindowOnError();
+      this.console.overrideWindowOnError();
       assert.notDeepStrictEqual(this.oldWindowOnError, window.oldWindowOnError);
     });
     it('should not break the existing window.onerror', function() {
-      console.restoreWindowOnError();
+      this.console.restoreWindowOnError();
       const initialCallCount = window.onErrorCallCount;
-      console.overrideWindowOnError();
+      this.console.overrideWindowOnError();
       window.onerror();
       assert.deepStrictEqual(window.onErrorCallCount, initialCallCount + 1);
     });
     it('should write the error to stderr', function() {
-      console.restoreWindowOnError();
-      console.overrideWindowOnError();
+      this.console.restoreWindowOnError();
+      this.console.overrideWindowOnError();
       stderr.inspectSync((output) => {
         window.onerror('', '', 0, 0, 'string');
         window.onerror('', '', 0, 0, [1, 2, 3]);
@@ -119,7 +121,7 @@ describe('Console', function() {
     it('should output the input and the evaluated object reference', function() {
       window.testString = 'value';
       stdout.inspectSync((output) => {
-        console.exec('window.testString');
+        this.console.exec('window.testString');
         assert.deepStrictEqual(output, [
           'window.testString\n',
           `${window.testString}\n`
@@ -128,20 +130,20 @@ describe('Console', function() {
     });
     it('should assign variables', function() {
       window.testString = 'old';
-      console.exec('window.testString = "new"');
+      this.console.exec('window.testString = "new"');
       assert.strictEqual(window.testString, 'new');
     });
     it('should not execute functions', function() {
       let testVar = 'old';
       window.testFn = () => (testVar = 'new');
-      console.exec('window.testFn()');
+      this.console.exec('window.testFn()');
       assert.strictEqual(testVar, 'old');
     });
-    it('should catch errors and console.error it out', function() {
+    it('should catch errors and this.console.error it out', function() {
       stderr.inspectSync((output) => {
         assert.deepStrictEqual(output.length, 0, this.setupError);
         const notAString = 123;
-        console.exec(notAString);
+        this.console.exec(notAString);
         assert.deepStrictEqual(output.length, 1);
       });
     });
@@ -151,7 +153,7 @@ describe('Console', function() {
     before(function() {
       this.testParseCommand = (command) => {
         assert.strictEqual(
-          console.parseCommand(command),
+          this.console.parseCommand(command),
           new Function(`return ${command}`)() // eslint-disable-line no-new-func
         );
       };
