+++ src/components/content/console/Output.test.js
