+++ src/components/content/console/Prompt.js
